Refactor store query implementation and documentation
