+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/query/DeleteStoreQueryRuntime.java
@@ -19,58 +19,20 @@ package org.wso2.siddhi.core.query;
 
 import org.wso2.siddhi.core.event.ComplexEvent;
 import org.wso2.siddhi.core.event.ComplexEventChunk;
-import org.wso2.siddhi.core.event.Event;
 import org.wso2.siddhi.core.event.state.StateEvent;
-import org.wso2.siddhi.core.event.state.StateEventPool;
 import org.wso2.siddhi.core.event.stream.MetaStreamEvent;
 import org.wso2.siddhi.core.event.stream.StreamEvent;
-import org.wso2.siddhi.core.exception.StoreQueryRuntimeException;
-import org.wso2.siddhi.core.query.selector.QuerySelector;
-import org.wso2.siddhi.query.api.definition.Attribute;
-
-import java.util.Arrays;
-import java.util.List;
 
 /**
  * This class is used to keep the runtime information needed to execute an delete store query.
  */
-public class DeleteStoreQueryRuntime implements StoreQueryRuntime {
-
-    private String queryName;
-    private MetaStreamEvent.EventType eventType;
-    private QuerySelector selector;
-    private StateEventPool stateEventPool;
-    private MetaStreamEvent metaStreamEvent;
-    private Attribute[] outputAttributes;
+public class DeleteStoreQueryRuntime extends StoreQueryRuntime {
 
     public DeleteStoreQueryRuntime(String queryName, MetaStreamEvent metaStreamEvent) {
         this.queryName = queryName;
         this.eventType = metaStreamEvent.getEventType();
         this.metaStreamEvent = metaStreamEvent;
         this.setOutputAttributes(metaStreamEvent.getLastInputDefinition().getAttributeList());
-    }
-
-    @Override
-    public Event[] execute() {
-        try {
-            StateEvent stateEvent = new StateEvent(1, outputAttributes.length);
-            StreamEvent streamEvent = new StreamEvent(metaStreamEvent.getBeforeWindowData().size(),
-                    metaStreamEvent.getOnAfterWindowData().size(),
-                    metaStreamEvent.getOutputData().size());
-            stateEvent.addEvent(0, streamEvent);
-
-            ComplexEventChunk complexEventChunk = new ComplexEventChunk(stateEvent, stateEvent, true);
-
-            if (eventType == MetaStreamEvent.EventType.TABLE) {
-                selector.process(complexEventChunk);
-            } else {
-                throw new StoreQueryRuntimeException("Delete store query consumes only stream events of type " +
-                        "\"TABLE\".");
-            }
-            return new Event[]{};
-        } catch (Throwable t) {
-            throw new StoreQueryRuntimeException("Error executing '" + queryName + "', " + t.getMessage(), t);
-        }
     }
 
     @Override
@@ -83,33 +45,6 @@ public class DeleteStoreQueryRuntime implements StoreQueryRuntime {
     @Override
     public TYPE getType() {
         return TYPE.DELETE;
-    }
-
-    public void setStateEventPool(StateEventPool stateEventPool) {
-        this.stateEventPool = stateEventPool;
-    }
-
-    /**
-     * This method sets selector for the delete store query runtime.
-     *
-     * @param selector for the store query
-     */
-    public void setSelector(QuerySelector selector) {
-        this.selector = selector;
-    }
-
-    /**
-     * This method sets the output attribute list of the given store query.
-     *
-     * @param outputAttributeList of the store query
-     */
-    public void setOutputAttributes(List<Attribute> outputAttributeList) {
-        this.outputAttributes = outputAttributeList.toArray(new Attribute[outputAttributeList.size()]);
-    }
-
-    @Override
-    public Attribute[] getStoreQueryOutputAttributes() {
-        return Arrays.copyOf(outputAttributes, outputAttributes.length);
     }
 
     private ComplexEventChunk<ComplexEvent> generateResetComplexEventChunk(MetaStreamEvent metaStreamEvent) {
