+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/query/FindStoreQueryRuntime.java
@@ -22,35 +22,26 @@ import org.wso2.siddhi.core.event.ComplexEvent;
 import org.wso2.siddhi.core.event.ComplexEventChunk;
 import org.wso2.siddhi.core.event.Event;
 import org.wso2.siddhi.core.event.state.StateEvent;
-import org.wso2.siddhi.core.event.state.StateEventPool;
 import org.wso2.siddhi.core.event.stream.MetaStreamEvent;
 import org.wso2.siddhi.core.event.stream.StreamEvent;
 import org.wso2.siddhi.core.exception.StoreQueryRuntimeException;
-import org.wso2.siddhi.core.query.selector.QuerySelector;
 import org.wso2.siddhi.core.table.Table;
 import org.wso2.siddhi.core.util.collection.operator.CompiledCondition;
 import org.wso2.siddhi.core.window.Window;
-import org.wso2.siddhi.query.api.definition.Attribute;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 /**
  * Store Query Runtime holds the runtime information needed for executing the store query.
  */
-public class FindStoreQueryRuntime implements StoreQueryRuntime {
+public class FindStoreQueryRuntime extends StoreQueryRuntime {
 
     private CompiledCondition compiledCondition;
     private Table table;
     private Window window;
-    private String queryName;
     private MetaStreamEvent.EventType eventType;
-    private AggregationRuntime aggregation;
-    private QuerySelector selector;
-    private StateEventPool stateEventPool;
-    private MetaStreamEvent metaStreamEvent;
-    private Attribute[] outputAttributes;
+    private AggregationRuntime aggregation;;
 
     public FindStoreQueryRuntime(Table table, CompiledCondition compiledCondition, String queryName,
                                  MetaStreamEvent metaStreamEvent) {
@@ -150,28 +141,6 @@ public class FindStoreQueryRuntime implements StoreQueryRuntime {
         ComplexEventChunk<ComplexEvent> complexEventChunk = new ComplexEventChunk<>(true);
         complexEventChunk.add(stateEvent);
         return complexEventChunk;
-    }
-
-    public void setStateEventPool(StateEventPool stateEventPool) {
-        this.stateEventPool = stateEventPool;
-    }
-
-    public void setSelector(QuerySelector selector) {
-        this.selector = selector;
-    }
-
-    /**
-     * This method sets the output attribute list of the given store query.
-     *
-     * @param outputAttributeList
-     */
-    public void setOutputAttributes(List<Attribute> outputAttributeList) {
-        this.outputAttributes = outputAttributeList.toArray(new Attribute[outputAttributeList.size()]);
-    }
-
-    @Override
-    public Attribute[] getStoreQueryOutputAttributes() {
-        return Arrays.copyOf(outputAttributes, outputAttributes.length);
     }
 
     private Event[] executeSelector(StreamEvent streamEvents, MetaStreamEvent.EventType eventType) {
