+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/query/UpdateStoreQueryRuntime.java
@@ -19,58 +19,20 @@ package org.wso2.siddhi.core.query;
 
 import org.wso2.siddhi.core.event.ComplexEvent;
 import org.wso2.siddhi.core.event.ComplexEventChunk;
-import org.wso2.siddhi.core.event.Event;
 import org.wso2.siddhi.core.event.state.StateEvent;
-import org.wso2.siddhi.core.event.state.StateEventPool;
 import org.wso2.siddhi.core.event.stream.MetaStreamEvent;
 import org.wso2.siddhi.core.event.stream.StreamEvent;
-import org.wso2.siddhi.core.exception.StoreQueryRuntimeException;
-import org.wso2.siddhi.core.query.selector.QuerySelector;
-import org.wso2.siddhi.query.api.definition.Attribute;
-
-import java.util.Arrays;
-import java.util.List;
 
 /**
  * This class is used to keep the runtime information needed to execute an update store query.
  */
-public class UpdateStoreQueryRuntime implements StoreQueryRuntime {
-
-    private String queryName;
-    private MetaStreamEvent.EventType eventType;
-    private QuerySelector selector;
-    private StateEventPool stateEventPool;
-    private MetaStreamEvent metaStreamEvent;
-    private Attribute[] outputAttributes;
+public class UpdateStoreQueryRuntime extends StoreQueryRuntime {
 
     public UpdateStoreQueryRuntime(String queryName, MetaStreamEvent metaStreamEvent) {
         this.queryName = queryName;
         this.eventType = metaStreamEvent.getEventType();
         this.metaStreamEvent = metaStreamEvent;
         this.setOutputAttributes(metaStreamEvent.getLastInputDefinition().getAttributeList());
-    }
-
-    @Override
-    public Event[] execute() {
-        try {
-            StateEvent stateEvent = new StateEvent(1, outputAttributes.length);
-            StreamEvent streamEvent = new StreamEvent(metaStreamEvent.getBeforeWindowData().size(),
-                    metaStreamEvent.getOnAfterWindowData().size(),
-                    metaStreamEvent.getOutputData().size());
-
-            stateEvent.addEvent(0, streamEvent);
-            ComplexEventChunk complexEventChunk = new ComplexEventChunk(stateEvent, stateEvent, true);
-
-            if (eventType == MetaStreamEvent.EventType.TABLE) {
-                selector.process(complexEventChunk);
-            } else {
-                throw new StoreQueryRuntimeException("Update store query consumes only stream events of type " +
-                        "\"TABLE\".");
-            }
-            return new Event[]{};
-        } catch (Throwable t) {
-            throw new StoreQueryRuntimeException("Error executing '" + queryName + "', " + t.getMessage(), t);
-        }
     }
 
     @Override
@@ -83,38 +45,6 @@ public class UpdateStoreQueryRuntime implements StoreQueryRuntime {
     @Override
     public TYPE getType() {
         return TYPE.UPDATE;
-    }
-
-    /**
-     * This method sets a state event pool for the update store query runtime.
-     *
-     * @param stateEventPool stateEventPool for the store query runtime
-     */
-    public void setStateEventPool(StateEventPool stateEventPool) {
-        this.stateEventPool = stateEventPool;
-    }
-
-    /**
-     * This method sets selector for the update query runtime.
-     *
-     * @param selector for the store query
-     */
-    public void setSelector(QuerySelector selector) {
-        this.selector = selector;
-    }
-
-    /**
-     * This method sets the output attribute list of the given store query.
-     *
-     * @param outputAttributeList output attributes of the store query
-     */
-    public void setOutputAttributes(List<Attribute> outputAttributeList) {
-        this.outputAttributes = outputAttributeList.toArray(new Attribute[outputAttributeList.size()]);
-    }
-
-    @Override
-    public Attribute[] getStoreQueryOutputAttributes() {
-        return Arrays.copyOf(outputAttributes, outputAttributes.length);
     }
 
     private ComplexEventChunk<ComplexEvent> generateResetComplexEventChunk(MetaStreamEvent metaStreamEvent) {
